requirements.txt
# Web UI
streamlit>=1.35.0

# LLM
anthropic>=0.28.0

# RAG
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Law API / HTML parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Scheduler
apscheduler>=3.10.0

# DB
# sqlite3 is built-in

# Auth
pyotp>=2.9.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
cryptography>=42.0.0

# PDF
reportlab>=4.0.0
xhtml2pdf>=0.2.16

# Session cookies
extra-streamlit-components>=0.1.60

# Utils
python-dotenv>=1.0.0
